+++ core/dbt/include/global_project/macros/schema_tests/accepted_values.sql
@@ -36,8 +36,7 @@ from validation_errors
 
 {% endmacro %}
 
-
-{% macro test_accepted_values(model, values) %}
+{% test accepted_values(model, values) %}
     {% set macro = adapter.dispatch('test_accepted_values') %}
     {{ macro(model, values, **kwargs) }}
-{% endmacro %}
+{% endtest %}